LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-anchor-fragment-expected.txt


FAIL Fragment <a> click fires navigate event assert_equals: expected (object) null but got (string) ""
